[standalone]: Move PLUGIN_NOTIFICATIONS_SDK slot to metabase/plugins/oss
